Finished Binary_Search_Tree.c

+++ Binary_Search_Tree.c
@@ -72,9 +72,26 @@ struct node* get_largest_element(struct node *root, struct node *largest)
     {
         largest = root;
     }
-    largest = get_largest_element(root->left, largest);
     largest = get_largest_element(root->right, largest);
-    return largest;
+    largest = get_largest_element(root->left, largest);
+}
+
+struct node* get_smallest_element(struct node *root, struct node *smallest)
+{
+    if(root == NULL)
+    {
+        return smallest;
+    }
+    if(smallest == NULL)
+    {
+        smallest = root;
+    }
+    else if(smallest->data >= root->data)
+    {
+        smallest = root;
+    }
+    smallest = get_smallest_element(root->left, smallest);   
+    smallest = get_smallest_element(root->right, smallest);
 }
 
 void delete_node(struct node *root, int number)
@@ -82,10 +99,100 @@ void delete_node(struct node *root, int number)
     struct node *temp1 = get_location(root, number);
     struct node *temp2 = get_largest_element(temp1->left, NULL);
     struct node *trash = NULL;
-    if(temp2->left == NULL && temp2->right == NULL)
+    if(temp2 == NULL)
+        temp2 = get_smallest_element(temp1->right, NULL);
+    if(temp2 == NULL)
     {
-        temp2->left = temp1->left;
-        temp2->right = temp1->right;
+        trash = temp1;
+        if((temp1->previous)->left == temp1)
+        {
+            (temp1->previous)->left = NULL;
+        }
+        else 
+        {
+            (temp1->previous)->right = NULL;
+        }
+        free(trash);
+    }
+    else if(temp2->left == NULL && temp2->right == NULL)
+    {
+        if(temp1->left != temp2)
+            temp2->left = temp1->left;
+        else
+            temp2->left = NULL;
+        if(temp1->right != temp2)
+            temp2->right = temp1->right;
+        else
+            temp2->right = NULL;
+        trash = temp1;
+        printf("%d\n", temp2->data);
+        if((temp1->previous)->left == temp1)
+        {
+            (temp1->previous)->left = temp2;
+            if((temp2->previous)->left == temp2)
+            {
+                (temp2->previous)->left = NULL;
+            }
+            else
+            {
+                (temp2->previous)->right = NULL;
+            }
+            temp2->previous = temp1->previous;
+        }
+        else if((temp1->previous)->right == temp1)
+        {
+            (temp1->previous)->right = temp2;
+            if((temp2->previous)->left == temp2)
+            {
+                (temp2->previous)->left = NULL;
+            }
+            else
+            {
+                (temp2->previous)->right = NULL;
+            }
+            temp2->previous = temp1->previous;
+        }
+        free(trash);
+    }
+    else 
+    {
+        struct node *new_temp1 = get_largest_element(temp2->left, NULL);
+        if(new_temp1 == NULL)
+            new_temp1 = get_smallest_element(temp2->right, NULL);
+        if(temp2->left != new_temp1)
+            new_temp1->left = temp2->left;
+        else
+            new_temp1->left = NULL;
+        if(temp2->right != new_temp1)
+            new_temp1->right = temp2->right;
+        else
+            new_temp1->right = NULL;
+        if((temp2->previous)->left == temp2)
+        {
+            (temp2->previous)->left = new_temp1;
+            if((new_temp1->previous)->left == new_temp1)
+                (new_temp1->previous)->left = NULL;
+            else
+                (new_temp1->previous)->right = NULL;
+            new_temp1->previous = temp2->previous;
+        }
+        else if((temp2->previous)->right == temp2)
+        {
+            (temp2->previous)->right = new_temp1;
+            if((new_temp1->previous)->left == new_temp1)
+                (new_temp1->previous)->left = NULL;
+            else
+                (new_temp1->previous)->right = NULL;
+            new_temp1->previous = temp2->previous;
+        }
+        if(temp1->left != temp2)
+            temp2->left = temp1->left;
+        else
+            temp2->left = NULL;
+        if(temp1->right != temp2)
+            temp2->right = temp1->right;
+        else
+            temp2->right = NULL;
         trash = temp1;
         if((temp1->previous)->left == temp1)
         {
@@ -98,8 +205,9 @@ void delete_node(struct node *root, int number)
             {
                 (temp2->previous)->right = NULL;
             }
+            temp2->previous = temp1->previous;
         }
-        else 
+        else if((temp1->previous)->right == temp1)
         {
             (temp1->previous)->right = temp2;
             if((temp2->previous)->left == temp2)
@@ -110,10 +218,10 @@ void delete_node(struct node *root, int number)
             {
                 (temp2->previous)->right = NULL;
             }
+            temp2->previous = temp1->previous;
         }
         free(trash);
     }
-
 }
 
 void display_tree(struct node *root)
@@ -161,7 +269,13 @@ int main()
                 {
                     printf("Enter the number to be deleted: ");
                     scanf("%d", &number);
-                    delete_node(root, number);
+                    if(root->data == number && root->left == NULL && root->right == NULL)
+                    {
+                        free(root);
+                        root = NULL;
+                    }
+                    else
+                        delete_node(root, number);
                 }
                 break;
